+++ src/Spinner.Test/Helper/Interceptors/InvalidInterceptor.cs
@@ -1,0 +1,8 @@
+namespace Spinner.Test.Helper.Interceptors
+{
+
+    public class InvalidInterceptor
+    {
+        public string Parse(string value) => value;
+    }
+}
