+++ src/Spinner.Test/Helper/Models/ModelWithInvalidInterceptor.cs
@@ -1,0 +1,11 @@
+using Spinner.Attribute;
+using Spinner.Test.Helper.Interceptors;
+
+namespace Spinner.Test.Helper.Models
+{
+    public class ModelWithInvalidInterceptor
+    {
+        [ReadProperty(0, 10, typeof(InvalidInterceptor))]
+        public string Value { get; set; }
+    }
+}
