+++ src/Spinner.Test/Helper/Models/ModelWithLargeLength.cs
@@ -1,0 +1,11 @@
+using Spinner.Attribute;
+
+namespace Spinner.Test.Helper.Models
+{
+    [ObjectMapper(300)]
+    public class ModelWithLargeLength
+    {
+        [WriteProperty(300, 0, ' ')]
+        public string Data { get; set; }
+    }
+}
