+++ src/Spinner.Test/Helper/Models/ModelWithReadOnlyPropertyForWrite.cs
@@ -1,0 +1,10 @@
+using Spinner.Attribute;
+
+namespace Spinner.Test.Helper.Models
+{
+    public class ModelWithReadOnlyPropertyForWrite
+    {
+        [ReadProperty(0, 10)]
+        public string Value => "ReadOnly";
+    }
+}
