+++ src/Spinner.Test/Helper/Models/ModelWithSmall50Length.cs
@@ -1,0 +1,14 @@
+using Spinner.Attribute;
+
+namespace Spinner.Test
+{
+    public partial class SpinnerWriteTest
+    {
+        [ObjectMapper(50)]
+        public class ModelWithSmall50Length
+        {
+            [WriteProperty(50, 0, ' ')]
+            public string Data { get; set; }
+        }
+    }
+}
