+++ src/Spinner.Test/Helper/Models/ModelWithZeroLength.cs
@@ -1,0 +1,11 @@
+using Spinner.Attribute;
+
+namespace Spinner.Test.Helper.Models
+{
+    [ObjectMapper(0)]
+    public class ModelWithZeroLength
+    {
+        [WriteProperty(10, 0, ' ')]
+        public string Value { get; set; }
+    }
+}
