+++ src/Spinner.Test/Helper/Models/ModelWithoutObjectMapperAttribute.cs
@@ -1,0 +1,10 @@
+using Spinner.Attribute;
+
+namespace Spinner.Test.Helper.Models
+{
+    public class ModelWithoutObjectMapperAttribute
+    {
+        [WriteProperty(10, 0, ' ')]
+        public string Value { get; set; }
+    }
+}
