+++ src/Spinner.Test/Helper/Models/NothingNoAttibute.cs
@@ -1,7 +1,7 @@
 ﻿using Spinner.Attribute;
 using System;
 
-namespace Spinner.Test.Models
+namespace Spinner.Test.Helper.Models
 {
     [ObjectMapper(length: 50)]
     internal sealed class NothingNoAttribute : IEquatable<NothingNoAttribute>
