+++ src/Spinner.Test/Helper/Models/NothingReader.cs
@@ -1,7 +1,7 @@
 ﻿using Spinner.Attribute;
 using System;
 
-namespace Spinner.Test.Models
+namespace Spinner.Test.Helper.Models
 {
     [ObjectMapper(length: 50)]
     internal sealed class NothingReader : IEquatable<NothingReader>
